renamed spill, event, event_model to match class names
