+++ .gitignore
@@ -1,4 +1,4 @@
-/build
+build/
 /install
 /source/config/*
 *.app
