+++ source/gui/daq/ConsumerDialog.h
@@ -7,7 +7,7 @@
 
 //#include "widget_detectors.h"
 #include <core/consumer.h>
-#include <core/spill.h>
+#include <core/Spill.h>
 
 namespace Ui
 {
