+++ source/producers/MockProducer/ValueDefinition.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <core/plugin/plugin.h>
-#include <core/event.h>
+#include <core/Event.h>
 #include <random>
 
 using namespace DAQuiri;
