+++ tests/core/EventModel.cpp
@@ -1,4 +1,4 @@
-#include <core/event_model.h>
+#include <core/EventModel.h>
 #include <gtest/gtest.h>
 
 TEST(EventModel, AddValue)
